requirements.txt
absl-py==2.0.0

asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
crc32c==2.3.post0
crcmod==1.7
cycler==0.12.1
decorator==4.4.2
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.12.1
gast==0.5.4
google-auth==2.25.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
imageio==2.33.0
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lxml==4.9.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
#nvidia-cudnn-cu12==8.9.2.26
#nvidia-cufft-cu12==11.0.2.54
#nvidia-curand-cu12==10.3.2.106
#nvidia-cusolver-cu12==11.4.5.107
#nvidia-cusparse-cu12==12.1.0.106
#nvidia-nccl-cu12==2.18.1
#nvidia-nvjitlink-cu12==12.3.101
#nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
osrf-pycommon==0.1.9
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
pluggy==1.3.0
proglog==0.1.10
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyserial==3.5
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
#quality-of-service-demo-py==0.7.9
requests==2.31.0
requests-oauthlib==1.3.1

scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
#teleop-twist-keyboard==2.3.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
#tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf-slim==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
#topic-monitor==0.7.9
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.14.0
#triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1


